+++ MusicalCollection/Song.h
@@ -13,18 +13,18 @@ private:
 	std::string genre;
 	std::string album;
 	int year;
-	double rating;	
+	float rating;	
 
 public:
 	Song(std::string _name, std::string _singer, std::string _genre, std::string _album, 
-		int _year, double _rating);
+		int _year, float _rating);
 
 	std::string getName();
 	std::string getSinger();
 	std::string getGenre();
 	std::string getAlbum();
 	Date getYear();
-	double getRating();
+	float getRating();
 	bool operator < (const Song& s) const;
 	friend std::ostream& operator << (std::ostream& output, const Song& song);
 };
